requirements.txt
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
